Split HIR reachability helpers
Move stdlib reachability and call-graph discovery out of the main HIR lowering facade into a tracked compiler.hir child module.

Lower the compiler source monolith ratchet for hir.rs, top-file lines, and top-file share; add reachability.rs to the boundary map and report fixture.

Refs #1254.
