+++ scripts/ci/report-compiler-source-monoliths.py
@@ -42,6 +42,7 @@ BOUNDARY_MAP: dict[str, list[str]] = {
     "new_project.rs": ["compiler.commands"],
     "ownership.rs": ["compiler.hir"],
     "properties.rs": ["compiler.hir"],
+    "reachability.rs": ["compiler.hir"],
     "project.rs": ["compiler.package_graph", "compiler.commands", "compiler.evidence"],
     "registry.rs": ["compiler.package_graph"],
     "signatures.rs": ["compiler.hir"],
